refactor: extract form elements in new component to reuse in both add and edit goal page.
