+++ src/pages/Goals/Goals.tsx
@@ -44,7 +44,7 @@ export default function Goals() {
   }, [data, isLoading])
 
   const getPercentage = useCallback((value: number, total: number) => {
-    return +((value / total) * 100).toFixed(2)
+    return +((value / total) * 100).toFixed(2) || 0
   }, [])
 
   if (isLoading) {
@@ -104,16 +104,25 @@ export default function Goals() {
       <div className='text-2xl font-bold truncate block py-4'>Your ongoing financial goals</div>
 
       <div className='grid sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4 gap-4 gap-y-10'>
-        {data?.map((goal) => (
-          <Goal
-            key={goal.id}
-            id={goal.id}
-            goalTitle={goal.name}
-            category={goal.track_kpi}
-            createdAt={dayjs(goal.created_at).toISOString()}
-            color={colors.violet['500']}
-          />
-        ))}
+        {data?.length ? (
+          data?.map((goal) => (
+            <Goal
+              key={goal.id}
+              id={goal.id}
+              goalTitle={goal.name}
+              category={goal.track_kpi}
+              createdAt={dayjs(goal.created_at).toISOString()}
+              color={colors.violet['500']}
+            />
+          ))
+        ) : (
+          <div>
+            <p>You Have no financial goals.</p>
+            <Link to='/goals/new' className='block underline mt-2'>
+              Click here to add.
+            </Link>{' '}
+          </div>
+        )}
       </div>
     </Page>
   )
